aider: Add logging of submission object in `testReviewerVote` function.
# Aider chat conversation:

USER: in `testReviewerVote` log submission object

ASSISTANT: To log the submission object in the `testReviewerVote` function, we need to modify the `test/PeerReview.t.sol` file. Specifically, we will add logging statements after each vote is cast to log the state of the submission object.

Here is the *SEARCH/REPLACE* block to implement this change:

test/PeerReview.t.sol
```solidity
<<<<<<< SEARCH
        vm.prank(reviewerAddresses[0]);
        peerReview.reviewerVote(1, submissionId);
        vm.prank(reviewerAddresses[1]);
        peerReview.reviewerVote(1, submissionId);
        vm.prank(reviewerAddresses[2]);
        peerReview.reviewerVote(1, submissionId);
=======
        vm.prank(reviewerAddresses[0]);
        peerReview.reviewerVote(1, submissionId);
        (address author0, , , , , uint256 countVotes0) = peerReview.getSubmission(submissionId);
        console.log("After 1st vote: Author:", author0, "CountVotes:", countVotes0);

        vm.prank(reviewerAddresses[1]);
        peerReview.reviewerVote(1, submissionId);
        (address author1, , , , , uint256 countVotes1) = peerReview.getSubmission(submissionId);
        console.log("After 2nd vote: Author:", author1, "CountVotes:", countVotes1);

        vm.prank(reviewerAddresses[2]);
        peerReview.reviewerVote(1, submissionId);
        (address author2, , , , , uint256 countVotes2) = peerReview.getSubmission(submissionId);
        console.log("After 3rd vote: Author:", author2, "CountVotes:", countVotes2);
>>>>>>> REPLACE
```

This change adds logging after each vote is cast, showing the author of the submission and the current count of votes. This will help in debugging and understanding the state of the submission object after each voting action.